+++ eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/Generator/Core.hs
@@ -71,7 +71,7 @@ import Cardano.Ledger.Credential (
  )
 import Cardano.Ledger.Hashes (unsafeMakeSafeHash)
 import Cardano.Ledger.Keys (VKey, asWitness)
-import Cardano.Ledger.Shelley.LedgerState (ChainAccountState (..))
+import Cardano.Ledger.Shelley.State
 import Cardano.Ledger.Shelley.TxWits (ShelleyTxWits)
 import Cardano.Ledger.Slot (
   Duration (..),
@@ -79,7 +79,6 @@ import Cardano.Ledger.Slot (
   epochInfoFirst,
   (*-),
  )
-import Cardano.Ledger.State (UTxO (UTxO))
 import Cardano.Ledger.TxIn (TxId, TxIn (TxIn))
 import Cardano.Protocol.Crypto (Crypto)
 import Cardano.Protocol.TPraos.OCert (KESPeriod (..))